requirements.txt
django==4.0.4
Pillow==8.4.0
mysqlclient==2.1.0
django-crispy-forms==1.13.0
djangorestframework==3.13.1
model-mommy==2.0.0
coverage==6.2
crispy-forms-foundation==0.8.0
django_widgets_improved==1.5.0
model-mommy==2.0.0
coverage==6.2
bcrypt==3.2.0
timeago==1.0.15
channels==3.0.4
channels-redis==3.4.0
python-dotenv==0.20.0
requests
python-dotenv
discord==1.7.3